refactor: isolate components and logic
