+++ src/state/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
-import { pokemonApi } from '../services/pokemon';
+import { pokemonApi } from './services/pokemon';
 
 export const store = configureStore({
   reducer: {
